Statically link logic library done, tests still working :).
