+++ src/Makefile.am
@@ -1,7 +1,7 @@
 AM_LDFLAGS=-static
 
-bin_PROGRAMS=mjs4
-mjs4_SOURCES = \
+lib_LTLIBRARIES = libmjs.la
+libmjs_la_SOURCES = \
 	engine/engine.c \
 	config/config.c \
 	songdata/dirstack.c \
@@ -16,6 +16,11 @@ mjs4_SOURCES = \
 	gui/playlist.c \
 	gui/info.c \
 	gui/playback.c \
-	gui/bar.c \
+	gui/bar.c
+libmjs_la_CFLAGS=$(apreq_includes) -std=c11 -Wall -Werror -D_GNU_SOURCE -D_LARGEFILE64_SOURCE
+
+bin_PROGRAMS=mjs4
+mjs4_SOURCES = \
 	mjs.c
 mjs4_CFLAGS=$(apreq_includes) -std=c11 -Wall -Werror -D_GNU_SOURCE -D_LARGEFILE64_SOURCE
+mjs4_LDADD=libmjs.la
