+++ src/gui/window.c
@@ -1,4 +1,4 @@
-#include "mjs.h"
+#include "utils.h"
 #include "window.h"
 #include "songdata/songdata.h"
 #include "config/config.h"
