+++ src/mjs.c
@@ -6,6 +6,7 @@
 #include "engine/engine.h"
 #include "config/config.h"
 #include "log.h"
+#include "utils.h"
 //#include "plugin/plugin.h"
 
 #include <string.h>
@@ -15,20 +16,11 @@
 #include <time.h>
 #include <locale.h>
 
-static struct sigaction handler;
 
 int main(int argc, char *argv[]) {
 	srand(time(NULL));
 
-	memset(&handler, 0, sizeof(struct sigaction));
-
-	/* Ignore the SIGQUIT signals a.k.a pressing Prt Scr */
-	handler.sa_handler = SIG_IGN;
-	handler.sa_flags = 0;
-	sigaction(SIGQUIT, &handler, NULL);
-
-	handler.sa_handler = (SIGHANDLER) bailout;
-	sigaction(SIGINT, &handler, NULL);
+	utils_init_sig_handlers();
 
 	//Prevent UTF-8 BUGS
 	setlocale(LC_ALL,"");
@@ -58,43 +50,4 @@ int main(int argc, char *argv[]) {
 	return 0;
 }
 
-void bailout(int sig) {	
-	/*
-	 * Lets get the hell out of here!
-	 */
-	handler.sa_handler = SIG_IGN;
-	handler.sa_flags = 0;
-	sigaction(SIGINT, &handler, NULL);
-	sigaction(SIGALRM, &handler, NULL);
 
-	switch (sig) {
-	default:
-		break;
-	}
-
-	//plugin_shutdown();
-	gui_shutdown();	
-	engine_shutdown();
-	songdata_shutdown();
-	controller_shutdown();
-	config_shutdown();
-	log_shutdown();
-
-	fprintf(stdout, "\n\n MP3 Jukebox System (mjs) v%s\n", VERSION);
-	fprintf(stdout, " Copyright (C) 2008\n");
-	fprintf(stdout,
-			" This program is free software; you can redistribute it and/or modify it\n");
-	fprintf(stdout,
-			" under the terms of the GNU General Public License as published by the Free\n");
-	fprintf(stdout, " Software Foundation; version 2 of the License.\n\n");
-	fprintf(
-			stdout,
-			" This program is distributed in the hope that it will be useful, but WITHOUT\n");
-	fprintf(stdout,
-			" ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or\n");
-	fprintf(stdout, " FITNESS FOR A PARTICULAR PURPOSE. \n\n");
-	fprintf(stdout, " See the GNU GPL (see LICENSE file) for more details.\n");
-	fprintf(stdout, " \n");
-
-	exit(sig);
-}
