+++ src/utils.h
@@ -2,6 +2,8 @@
 #define _utils_h
 
 
-char * strtrim(char *);
+char * strtrim( char * );
+void bailout( int );
+void utils_init_sig_handlers( void );
 
 #endif
